d/patches: Mark several additional tests as flaky
Most of these have never been reliable on Debian buildds, especially
on ARM, so having them fail is not a regression. They usually succeed
after a few retries, but that's time-consuming and unsuitable to be a
QA gate.

The restore-size test was previously not run at all: now it is run, but
marked as flaky (so its result is ignored).