d/rules, d/tests/ensure-x11-dir: Explicitly create /tmp/.X11-unix
mutter's test suite requires this directory, so create it explicitly
instead of relying on xvfb-run to have created it. If creating it
fails for whatever reason, this will at least give us better
diagnostic information.

We cannot create it with its canonical ownership (root:root) because
the build isn't running as root, but we can at least make sure it has
its canonical permissions.

Closes: #1084133