d/rules: Only run the test-suite for architecture-specific builds
Similar to other packages with large and/or unreliable test suites,
running the test suite once per architecture is enough: we don't need
to run it a second time (in practice on amd64) in distributions like
Debian that build Architecture: all separately.

In Ubuntu, which builds Architecture: all packages in the same run as
amd64, we'll still run the test suite during that amd64+all build.